fix compiler warnings

+++ branches/SHM-6-dev/source/Lib/TLibCommon/SEI.h
@@ -416,15 +416,15 @@ public:
   }
   virtual ~SEIColorMappingInfo() {}
 
-  Int    m_colorMapId;
-  Bool   m_colorMapCancelFlag;
-  Bool   m_colorMapPersistenceFlag;
-  Bool   m_colorMap_video_signal_type_present_flag;
-  Bool   m_colorMap_video_full_range_flag;
-  UShort m_colorMap_primaries;
-  UShort m_colorMap_transfer_characteristics;
-  UShort m_colorMap_matrix_coeffs;
-  UShort m_colorMapModelId;
+  Int   m_colorMapId;
+  Bool  m_colorMapCancelFlag;
+  Bool  m_colorMapPersistenceFlag;
+  Bool  m_colorMap_video_signal_type_present_flag;
+  Bool  m_colorMap_video_full_range_flag;
+  Int   m_colorMap_primaries;
+  Int   m_colorMap_transfer_characteristics;
+  Int   m_colorMap_matrix_coeffs;
+  Int   m_colorMapModelId;
 
   Int   m_colour_map_coded_data_bit_depth;
   Int   m_colour_map_target_bit_depth;
